requirements.txt
pygame
customtkinter
pillow
matplotlib
numpy
pandas
scikit-learn
